backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
webauthn==1.11.1
pyjwt==2.8.0
cryptography==41.0.7
geoip2==4.7.0
user-agents==2.2.0
httpx==0.25.2
pytz==2024.1
